Checking when removing the car in admin panel if the car is not currently rented

+++ src/main/java/com/app/carrent/controller/AdminController.java
@@ -211,7 +211,7 @@ public class AdminController {
 
         Optional<Car> carOptional = carService.findCarById(id);
         if (carOptional.isPresent()) {
-            actionOnCarItem(action, carOptional);
+            modelAndView = actionOnCarItem(action, modelAndView, carOptional);
         } else {
             modelAndView = new ModelAndView("forward:/admin/car-list-admin");
             modelAndView.addObject("carListAdminError", "Car with given id does not exist");
@@ -220,14 +220,21 @@ public class AdminController {
         return modelAndView;
     }
 
-    private void actionOnCarItem(@PathVariable("action") String action, Optional<Car> carOptional) {
+    private ModelAndView actionOnCarItem(@PathVariable("action") String action, ModelAndView modelAndView, Optional<Car> carOptional) {
         switch (action) {
             case "remove":
                 Car car = carOptional.get();
-                carService.delete(car);
+                Optional<CarRent> cr = carRentService.findCarRentItemByGivenCarWhereNowIsBetweenPickUpDateAndDropOffDate(LocalDateTime.now(), car);
+                if (cr.isPresent()) {
+                    modelAndView = new ModelAndView("forward:/admin/car-list-admin");
+                    modelAndView.addObject("carListAdminError", "You cannot delete the car because it is now rented");
+                } else
+                    carService.delete(car);
                 break;
         }
+        return modelAndView;
     }
+
 
     @GetMapping("/admin/car-list-admin/addCar")
     public ModelAndView addCar() {
